requirements.txt
streamlit
pandas
numpy
pyyaml
plotly
google-generativeai
